refactor: move client key handling
